gemini_chat/requirements.txt
google-genai>=0.8.0
python-dotenv>=1.0.0
httpx>=0.27.0
rich>=13.0.0
